% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iddconf23_2-package.R
\docType{package}
\name{iddconf23_2-package}
\alias{iddconf23_2}
\alias{iddconf23_2-package}
\title{iddconf23_2: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/p-ruiz-cuenca/iddconf23_2}
  \item Report bugs at \url{https://github.com/p-ruiz-cuenca/iddconf23_2/issues}
}

}
\author{
\strong{Maintainer}: Pablo Ruiz Cuenca \email{p.ruizcuenca@lancaster.ac.uk} (\href{https://orcid.org/0000-0002-2180-9509}{ORCID}) [copyright holder]

}
\keyword{internal}
